Inited draft training

+++ hicdifflib/utils.py
@@ -15,7 +15,7 @@ _unwanted_chars = re.compile(r'[^ACTG]')
 def sequence_to_onehot(sequence: str) -> _Tensor['nucleotide', 'sequence']:
     sequence = list(re.sub(_unwanted_chars, 'N', sequence.upper()))
     codes = ['ACTGN'.index(char) for char in sequence]
-    onehot = F.one_hot(torch.tensor(codes), 5).to(torch.float32)
+    onehot = F.one_hot(torch.tensor(codes, dtype=torch.long), 5).to(torch.float32)
     return torch.transpose(onehot, 0, 1)
